requirements.txt
numba
sympy
motmetrics==1.1.3
numpy
pandas>=0.24,<2.0
nuscenes-devkit==1.1.10
shapely
pyquaternion
line_profiler
lap
pyyaml

